+++ src/app/inventory/store/d2-item-factory.service.js
@@ -191,7 +191,7 @@ export function D2ItemFactory(
       isVendorItem: (!owner || owner.id === null),
       name: itemDef.displayProperties.name,
       description: itemDef.displayProperties.description,
-      icon: itemDef.displayProperties.icon,
+      icon: itemDef.displayProperties.icon || '/img/misc/missing_icon_d2.png',
       notransfer: Boolean(currentBucket.accountWide || currentBucket.inPostmaster || itemDef.nonTransferrable),
       id: item.itemInstanceId || '0', // zero for non-instanced is legacy hack
       equipped: Boolean(instanceDef.isEquipped),