+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added back in Repuation for D2.
 * Max Light Loadout, Make Room for Postmaster, Farming Mode, and Search Loadout are all reenabled for D2.
+* Classified items can be transferred.
 
 # 4.13.0