+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+# 4.14.0
+
 * Added back in Repuation for D2.
 * Max Light Loadout, Make Room for Postmaster, Farming Mode, and Search Loadout are all reenabled for D2.
 * Classified items can be transferred!